+++ crates/cubecl-convolution/src/tests/convolution_test_launcher.rs
@@ -12,7 +12,7 @@ use crate::base::ConvolutionProblem;
 use cubecl_matmul::components::InputIdent;
 use cubecl_matmul::components::global::args::{ConcreteOutputFactory, MatmulArgs};
 use cubecl_matmul::tests::test_utils::Sample;
-use cubecl_matmul::{components::Ident, tests::cmma_matmul::matmul_test_launcher::TensorRawParts};
+use cubecl_matmul::{components::Ident, tests::layered::matmul_test_launcher::TensorRawParts};
 
 use super::test_utils::TestPrecision;
