+++ crates/cubecl-matmul/src/tests/layered/macros/common/advanced/hypercube.rs
