+++ crates/cubecl-matmul/src/tests/layered/macros/plane_accelerated/algorithm.rs
