+++ crates/cubecl-matmul/src/tests/layered/mod.rs
@@ -1,2 +1,3 @@
+mod macros;
 pub mod matmul_test_launcher;
 pub mod tma_test_launcher;
