+++ crates/cubecl-matmul/src/tests/mod.rs
@@ -1,8 +1,5 @@
 #![allow(missing_docs)]
 
-pub mod cmma_matmul;
-pub mod simple;
-pub mod test_macros;
+pub mod layered;
+pub mod naive;
 pub mod test_utils;
-
-pub use test_utils::*;
