+++ crates/cubecl-matmul/src/tests/naive/mod.rs
@@ -1,0 +1,3 @@
+mod macros;
+pub mod tests;
+mod utils;
