--- crates/cubecl-matmul/src/tests/test_macros/mod.rs
@@ -1,2 +1,0 @@
-mod simple;
-pub mod suite;
